+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Feat: [iOS] Add support for call logging via `enableCallLogging(bool)` to record call in recents. No other platform currently supports this, see [NOTES.md](NOTES.md#limitations) for more details.
 * Fix: `showMissedCallNotifications` method call not working due to incorrect method channel name.
 * Fix: [Android] fix permission request callback not completing on permission re-request.
+* Fix: [Android] add missing Android Foreground Service `phoneCall`.
 * Feat: update example.
 * Docs: update CHANGELOG